ESPHome-SDL-UI_Beta/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pyyaml==6.0.2
aiofiles==24.1.0
pydantic==2.10.3
jinja2==3.1.4
python-dotenv==1.0.1
